+++ GUI/DataManager/DataManagerFrame.lua
@@ -5,20 +5,18 @@ local dataManagerFrame = addon.GUI.DataManagerFrame;
 
 function dataManagerFrame:Load()
 	local frame;
-	if addon.IsNotWotLKClassic() then
-		frame = CreateFrame("Frame", "KrowiAF_DataManagerFrame", UIParent, "KrowiAF_DataManagerFrame_Template");
-		KrowiAF_DataManagerFrameCloseButton:SetScript("OnKeyDown", function(selfFunc, key)
-			if key == GetBindingKey("TOGGLEGAMEMENU") then
-				if selfFunc:GetParent():IsShown() then
-					-- selfFunc:Click(selfFunc);
-					selfFunc:GetParent():Hide();
-					selfFunc:SetPropagateKeyboardInput(false);
-					return;
-				end
+	frame = CreateFrame("Frame", "KrowiAF_DataManagerFrame", UIParent, "KrowiAF_DataManagerFrame_Template");
+	KrowiAF_DataManagerFrameCloseButton:SetScript("OnKeyDown", function(selfFunc, key)
+		if key == GetBindingKey("TOGGLEGAMEMENU") then
+			if selfFunc:GetParent():IsShown() then
+				-- selfFunc:Click(selfFunc);
+				selfFunc:GetParent():Hide();
+				selfFunc:SetPropagateKeyboardInput(false);
+				return;
 			end
-			selfFunc:SetPropagateKeyboardInput(true);
-		end);
-	end
+		end
+		selfFunc:SetPropagateKeyboardInput(true);
+	end);
 
 	frame.ResetPosition = self.ResetPosition;
 	addon.GUI.DataManagerFrame = frame;