+++ GUI/GUI.lua
@@ -189,17 +189,15 @@ function gui.ToggleAchievementFrame(_addonName, tabName, resetView, forceOpen) -
     end
 
 	if AchievementFrame:IsShown() and tabIsSelected and not resetView and not forceOpen then
-		-- HideUIPanel(AchievementFrame);
         AchievementFrame:Hide();
 	else
-        if addon.IsNotWotLKClassic() then
+        if addon.IsNotWrathClassic() then
             AchievementFrame_SetTabs();
         else
             addon.GUI.ShowHideTabs();
         end
-		-- ShowUIPanel(AchievementFrame);
         AchievementFrame:Show();
-        if addon.IsNotWotLKClassic() then
+        if addon.IsNotWrathClassic() then
             AchievementFrame_HideSearchPreview();
         end
         gui.SelectTab(_addonName, tabName);
@@ -258,19 +256,19 @@ function gui.AddDataToBlizzardTabs()
     KrowiAF_RegisterTabButton("Blizzard_AchievementUI", "Achievements", AchievementFrameTab1, function()
         AchievementFrameTab_OnClick(1);
     end);
-    if addon.IsNotWotLKClassic() then
+    if addon.IsNotWrathClassic() then
         KrowiAF_RegisterTabButton("Blizzard_AchievementUI", "Guild", AchievementFrameTab2, function()
             AchievementFrameTab_OnClick(2);
         end);
     end
-    KrowiAF_RegisterTabButton("Blizzard_AchievementUI", "Statistics", addon.IsNotWotLKClassic() and AchievementFrameTab3 or AchievementFrameTab2, function()
-        AchievementFrameTab_OnClick(addon.IsNotWotLKClassic() and 3 or 2);
+    KrowiAF_RegisterTabButton("Blizzard_AchievementUI", "Statistics", addon.IsNotWrathClassic() and AchievementFrameTab3 or AchievementFrameTab2, function()
+        AchievementFrameTab_OnClick(addon.IsNotWrathClassic() and 3 or 2);
     end);
 end
 
 function gui.PrepareTabsOrder()
     KrowiAF_RegisterTabOptions("Blizzard_AchievementUI", "Achievements", addon.L["Blizzard"], addon.L["Achievements"], "TOGGLEACHIEVEMENT");
-    if addon.IsNotWotLKClassic() then
+    if addon.IsNotWrathClassic() then
         KrowiAF_RegisterTabOptions("Blizzard_AchievementUI", "Guild", addon.L["Blizzard"], addon.L["Guild"]);
     else
         addon.Options.Defaults.profile.Tabs.Blizzard_AchievementUI.Guild = nil;